cleaned tables_marker.py
